# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Organisation API","version":"0.1.0"},"paths":{"/v1/activities/{activity_id}/organisations":{"get":{"tags":["activities"],"summary":"Get Organisations","operationId":"get_organisations_v1_activities__activity_id__organisations_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"activity_id","in":"path","required":true,"schema":{"type":"integer","title":"Activity Id"}},{"name":"depth","in":"query","required":false,"schema":{"type":"integer","maximum":3,"minimum":1,"default":1,"title":"Depth"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganisationOut"},"title":"Response Get Organisations V1 Activities  Activity Id  Organisations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/buildings/{building_id}/organisations":{"get":{"tags":["buildings"],"summary":"Get Organisations","operationId":"get_organisations_v1_buildings__building_id__organisations_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"building_id","in":"path","required":true,"schema":{"type":"integer","title":"Building Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganisationOut"},"title":"Response Get Organisations V1 Buildings  Building Id  Organisations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/geo/organisations":{"get":{"tags":["geo"],"summary":"Get By Radius","operationId":"get_by_radius_v1_geo_organisations_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","maximum":90,"minimum":-90,"title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","maximum":180,"minimum":-180,"title":"Lon"}},{"name":"radius","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","exclusiveMinimum":0},{"type":"null"}],"description":"Radius in meters","title":"Radius"},"description":"Radius in meters"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganisationOut"},"title":"Response Get By Radius V1 Geo Organisations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/geo/rectangle/organisations":{"get":{"tags":["geo"],"summary":"Get By Rectangle","operationId":"get_by_rectangle_v1_geo_rectangle_organisations_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","maximum":90,"minimum":-90,"title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","maximum":180,"minimum":-180,"title":"Lon"}},{"name":"width","in":"query","required":true,"schema":{"anyOf":[{"type":"integer","exclusiveMinimum":0},{"type":"null"}],"description":"Width in meters","title":"Width"},"description":"Width in meters"},{"name":"length","in":"query","required":true,"schema":{"anyOf":[{"type":"integer","exclusiveMinimum":0},{"type":"null"}],"description":"Length in meters","title":"Length"},"description":"Length in meters"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganisationOut"},"title":"Response Get By Rectangle V1 Geo Rectangle Organisations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/organisations/":{"get":{"tags":["organisations"],"summary":"Get Organisations","operationId":"get_organisations_v1_organisations__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by name","title":"Name"},"description":"Search by name"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganisationOut"},"title":"Response Get Organisations V1 Organisations  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/organisations/{organisation_id}":{"get":{"tags":["organisations"],"summary":"Get Organisation","operationId":"get_organisation_v1_organisations__organisation_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"organisation_id","in":"path","required":true,"schema":{"type":"integer","title":"Organisation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganisationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActivityOut":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"integer","title":"Id"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"}},"type":"object","required":["name","id"],"title":"ActivityOut"},"BuildingOut":{"properties":{"address":{"type":"string","title":"Address"},"coordinates":{"$ref":"#/components/schemas/Coordinate","title":"Coordinates"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["address","coordinates","id"],"title":"BuildingOut"},"Coordinate":{"properties":{"latitude":{"anyOf":[{"type":"number","maximum":90.0,"minimum":-90.0},{"type":"string"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number","maximum":180.0,"minimum":-180.0},{"type":"string"}],"title":"Longitude"}},"type":"object","required":["latitude","longitude"],"title":"Coordinate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrganisationOut":{"properties":{"name":{"type":"string","title":"Name"},"phones":{"items":{"type":"string","format":"phone"},"type":"array","title":"Phones","default":[]},"id":{"type":"integer","title":"Id"},"building":{"$ref":"#/components/schemas/BuildingOut"},"activities":{"items":{"$ref":"#/components/schemas/ActivityOut"},"type":"array","title":"Activities","default":[]}},"type":"object","required":["name","id","building"],"title":"OrganisationOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}